remove planners
