+++ yucca/pipeline/planning/YuccaPlannerV2.py
@@ -10,7 +10,7 @@ class YuccaPlannerV2(YuccaPlanner):
 
     def determine_norm_op_per_modality(self):
         self.norm_op_per_modality = [
-            "ct" if "ct" in mod.lower() else "standardize" for mod in self.dataset_properties["modalities"].values()
+            "ct" if "ct" in mod.lower() else "volume_wise_znorm" for mod in self.dataset_properties["modalities"].values()
         ]
 
     def determine_transpose(self):
@@ -20,3 +20,17 @@ class YuccaPlannerV2(YuccaPlanner):
         sorting_key = np.argsort(median_spacing)[::-1]
         self.transpose_fw = sorting_key.tolist()
         self.transpose_bw = sorting_key.argsort().tolist()
+
+    def determine_target_size_from_fixed_size_or_spacing(self):
+        # If the dataset contains data with very different spacings this planner will
+        # not use the median spacing, but rather the first quartile spacing.
+        # By doing so it intentionally avoids excessive downsampling and instead favors upsampling.
+        self.fixed_target_size = None
+        self.fixed_target_spacing = self.dataset_properties["original_median_spacing"]
+        spacing_anisotropy_threshold = 2.5
+        spacings = np.array(self.dataset_properties["original_spacings"])
+        for dim in range(spacings.shape[1]):
+            low = np.percentile(spacings[:, dim], 5)
+            high = np.percentile(spacings[:, dim], 95)
+            if low * spacing_anisotropy_threshold < high:
+                self.fixed_target_spacing[dim] = np.percentile(spacings[:, dim], 10)
